Implementation for two new features
Added implementation to expose two new functionalities from the c++ to Torquescript as requested in issue #1272.  Added both a DefineEngineMethod function for getting the count of the number of objects on the turret ignore list and a DefineEngineMethod function for returning a reference to the object on the ignore list at a given index on the ignore list.  Also added functions that do each of those things and are then encapsulated by those DefineEngineMethod functions.